requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
Flask-Bcrypt
python-dotenv
PyMySQL